Add the feature "enum struct reflection."
FHT can now recognize the C++ enum structs.
With this feature, every enum field value can be converted to a string and vice versa.

+++ SDK/Header-Tool/Include/vocabulary.hpp
@@ -23,7 +23,7 @@ limitations under the License.
 
 
 
-enum struct Vocabulary : FE::uint16
+enum struct Vocabulary : FE::uint32
 {
 	_Undefined, _EndOfCode,
 	_LineComment, _LineEnd,
